Updated based on comments on PR 563
Splash page has been moved to the public unauthenticated space.
The protected space has been renamed 'secured' so that remix-flat-routes will provide the public spage first.
The layout route has been moved into the apply subfolder so that it does not impact the splash page.
Button classes updated so that layout better adjusts to smaller viewport sizes.
Name space passed to routeHandle function to address hyrdation error (and the accompaying white page flashing).
The src attribute is now a string where applicable.
Superfluous forced spaces removed.
Removed a redundant fragment with a single child.
